refactor(geodesy): rename regular_domain module
Rename regular_domain to regular_extent.
