+++ src/goesdl/geodesy/__init__.py
@@ -51,7 +51,7 @@ from .geos_to_latlon_grid_goesdl import geos_to_latlon_grid_goesdl
 from .geos_to_latlon_grid_pyproj import geos_to_latlon_grid_pyproj
 from .geostationary_parameters import GeostationaryParameters
 from .helpers import calculate_pixel_edges
-from .rectangular_domain import RectangularDomain
+from .rectangular_extent import RectangularDomain
 
 __all__ = [
     "geos_to_latlon_grid",
