+++ src/goesdl/geodesy/rectangular_extent.py
